Update jquery.form.js
With the jQuery versions higher than 1.8.3 (1.10.2 in our site), I was encountering an issue where an Ajax file upload was not being sent as multi-part. I found this issue was due to the value attribute selector not getting the correct value for the file input element, so [value!=''] was always false. I changed it to use the filter function instead and used val() to check the actual value. This works better at checking the value.