Add support for file download using cookie
This commit enable support for detecting a successful file download using cookie.
This feature is useful when exchanging files, upload a file a get a new one as result.
Since the plugin jQuery.fileDownload doesn't support well the file upload part, I choose to add this bit to this library.

Added the new option 'fileDownload:bool'.
When enabled a polling check is started that waits for the cookie 'jQueryFormFileDownload=true' to be set.

Signed-off-by: AndreaCuneo <andrea.cuneo@ark-energy.eu>